+++ internal/storage/storage_handle_test.go
@@ -1,4 +1,4 @@
-package gcsx
+package storage
 
 import (
 	"context"
